w3anno output: added --nested-key-delimiter paramter
Set this to a single character that is used to delimit keys that you
want to expand into multiple blank nodes/maps in the output. For
instance, set this to `.` and use it with keys such as `metadata.blah`
to get an extra grouping `metadata` in the body. If not set, nested keys
are not interpreted (default)")