fromxml: untangle multiple XML files into a single output text
Up until now stam fromxml assumed that each input XML file leads to one
output text file. This is not flexibile enough, we need to be able to
untangle multiple XML files to a single plain text resource. This is
customary for for example PageXML where each page is a separate XML
file.

Ref: https://github.com/annotation/stam-tools/issues/18